Refactor file/dir structure.
