+++ index.html
@@ -12,6 +12,6 @@
             <div id="root"></div>
         </div>
     </main>
-    <script type="module" src="/src/js/main.jsx"></script>
+    <script type="module" src="/src/js/index.jsx"></script>
 </body>
 </html>
