+++ src/js/Streams.jsx
@@ -3,7 +3,7 @@ import { useAutoAnimate } from "@formkit/auto-animate/react";
 import { TextBlur } from "./TextAnimations.jsx";
 import { selectAllStreams, selectStreamById } from "./services/qws/streams.js";
 import { useSelector } from "react-redux";
-import { pluralize } from "./util/text.js";
+import { pluralize } from "./common/text.js";
 
 export default function Streams() {
   const [parent] = useAutoAnimate();
