+++ src/js/app/App.jsx
@@ -1,11 +1,11 @@
 import React from "react";
-import ServerFilters from "./Servers/Filters.jsx";
-import ServerOverview from "./Servers/Overview.jsx";
-import Streams from "./Streams.jsx";
-import { streamsSlice } from "./services/qws/streams.js";
+import ServerFilters from "../servers/Filters.jsx";
+import ServerOverview from "../servers/Overview.jsx";
+import Streams from "../Streams.jsx";
+import { streamsSlice } from "../services/qws/streams.js";
 import store from "./store.js";
-import { serversSlice } from "./services/qws/servers.js";
-import Servers from "./Servers/Servers";
+import { serversSlice } from "../services/qws/servers.js";
+import Servers from "../servers/Servers.jsx";
 
 const AppHeader = () => {
   return (
