+++ src/js/common/storage.js
