+++ src/js/common/text.js
