+++ src/js/index.jsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import store from "./store.js";
+import store from "./app/store.js";
 import { Provider } from "react-redux";
-import App from "./App.jsx";
-import "./../styles/main.scss";
+import App from "./app/App.jsx";
+import "../styles/index.scss";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   // <React.StrictMode>
