+++ src/js/servers/FavoriteToggle.jsx
