+++ src/js/servers/ServerById.jsx
@@ -3,7 +3,7 @@ import FavoriteToggle from "./FavoriteToggle.jsx";
 import { Scoreboard } from "./Scoreboard.jsx";
 import { QuakeText } from "./QuakeText.jsx";
 import copyToClipboard from "copy-text-to-clipboard";
-import { pluralize } from "../util/text.js";
+import { pluralize } from "../common/text.js";
 import { TextBlur } from "../TextAnimations.jsx";
 import { useSelector } from "react-redux";
 import { selectServerById } from "../services/qws/servers.js";
